backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
python-jose[cryptography]